requirements.txt
Django==1.11.3
MySQL-python==1.2.5
django-solo==1.1.2
django-ckeditor==5.2.2